Fix regression browsing abilities in PC downtime/exploration and vehicles
