+++ static/templates/actors/character/tabs/actions.hbs
@@ -41,7 +41,7 @@
                                     <button type="button" class="item-control item-create" data-type="action" data-action-type="{{sid}}">
                                         <i class="fa-solid fa-fw fa-plus"></i>{{localize "PF2E.CreateActionTitle"}}
                                     </button>
-                                    <button type="button" class="item-control" data-action="browse-abilities" data-type="action" data-action-type="{{sid}}">
+                                    <button type="button" class="item-control" data-action="browse-abilities" data-action-type="{{sid}}">
                                         <i class="fa-solid fa-fw fa-search"></i>{{localize "PF2E.BrowseLabel"}}
                                     </button>
                                 </div>
@@ -84,7 +84,7 @@
                                 <button type="button" class="item-control item-create" data-type="action" data-action-type="free" data-traits="exploration">
                                     <i class="fa-solid fa-fw fa-plus"></i>{{localize "PF2E.CreateActionTitle"}}
                                 </button>
-                                <button type="button" class="item-control action-browse" data-type="action" data-action-trait="exploration">
+                                <button type="button" class="item-control" data-action="browse-abilities" data-action-trait="exploration">
                                     <i class="fa-solid fa-fw fa-search"></i>{{localize "PF2E.BrowseLabel"}}
                                 </button>
                             </div>
@@ -106,7 +106,7 @@
                                 <button type="button" class="item-control item-create" data-type="action" data-action-type="free" data-traits="downtime">
                                     <i class="fa-solid fa-fw fa-plus"></i>{{localize "PF2E.CreateActionTitle"}}
                                 </button>
-                                <button type="button" class="item-control action-browse" data-type="action" data-action-trait="downtime">
+                                <button type="button" class="item-control" data-action="browse-abilities" data-action-trait="downtime">
                                     <i class="fa-solid fa-fw fa-search"></i>{{localize "PF2E.BrowseLabel"}}
                                 </button>
                             </div>
